mise à jour

+++ src/main/java/org/psychohelp/psychohelp/repository/CitoyenRepository.java
@@ -3,5 +3,5 @@ package org.psychohelp.psychohelp.repository;
 import org.psychohelp.psychohelp.entity.Citoyen;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CitoyenRepository extends JpaRepository <Citoyen, Long> {
+public interface CitoyenRepository extends JpaRepository <Citoyen, Integer> {
 }
